backend/requirements.txt
fastapi
redis
types-redis
uvicorn
openai
python-dotenv
python-multipart